third_party/WebKit/LayoutTests/typedcssom/cssPerspective.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(() => {
  let percentValue = new CSSUnitValue(10, 'percent');
  let angleValue = new CSSUnitValue(20, 'deg');
  assert_throws(new TypeError(), () => { new CSSPerspective(percentValue) });
  assert_throws(new TypeError(), () => { new CSSPerspective(angleValue) });
}, "Constructor should throw an error for invalid CSSUnitValues");

test(() => {
  let perspective = new CSSPerspective(new CSSUnitValue(1, 'px'));
  assert_equals(perspective.length.value, 1);
  assert_equals(perspective.length.unit, 'px');

  perspective.length = new CSSUnitValue(2, 'in');
  assert_equals(perspective.length.value, 2);
  assert_equals(perspective.length.unit, 'in');
}, "Setting length with valid values.");

test(() => {
  let perspective = new CSSPerspective(new CSSUnitValue(1, 'px'));
  let percentUnit = new CSSUnitValue(2, 'percent');
  assert_throws(new TypeError(), () => { perspective.length = percentUnit; })
}, "Setting length with invalid values.");

test(() => {
  let perspective = new CSSPerspective(new CSSUnitValue(10, 'px'));

  assert_equals(perspective.toString(), 'perspective(10px)');
}, "toString should return perspective(<CSSNumericValue.cssString()>)");

</script>
